+++ pyrdf2vec/walkers/anonymous.py
@@ -12,20 +12,6 @@ class AnonymousWalker(RandomWalker):
     """Walker that transforms label information into positional information in
     order to anonymize the random walks.
 
-    Args:
-        max_depth: The maximum depth of one walk.
-        max_walks: The maximum number of walks per entity.
-        sampler: The sampling strategy.
-            Defaults to pyrdf2vec.samplers.UniformSampler().
-        n_jobs: The number of process to use for multiprocessing.
-            Defaults to 1.
-        with_reverse: extracts children's and parents' walks from the root,
-            creating (max_walks * max_walks) more walks of 2 * depth.
-            Defaults to False.
-        random_state: The random state to use to ensure ensure random
-            determinism to generate the same walks for entities.
-            Defaults to None.
-
     """
 
     def _extract(self, kg: KG, instance: Vertex) -> EntityWalks: